% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{heurist_fields}
\alias{heurist_fields}
\title{List Heurist Fields}
\usage{
heurist_fields(session)
}
\arguments{
\item{session}{A \code{heurist_session}.}
}
\value{
A list of Heurist field objects.
}
\description{
Retrieves field metadata for the current database.
}
