% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{heurist_logout}
\alias{heurist_logout}
\title{Log Out of Heurist}
\usage{
heurist_logout(session)
}
\arguments{
\item{session}{A \code{heurist_session}.}
}
\value{
The updated \code{heurist_session}.
}
\description{
Logs out an authenticated \code{heurist_session}.
}
