% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{heurist_raw_entity}
\alias{heurist_raw_entity}
\title{Call Heurist Entity Endpoint Directly}
\usage{
heurist_raw_entity(session, action, entity, query = list())
}
\arguments{
\item{session}{A \code{heurist_session}.}

\item{action}{Entity action.}

\item{entity}{Entity name.}

\item{query}{Additional query parameters.}
}
\value{
Parsed JSON response.
}
\description{
Low-level wrapper around \code{entityScrud.php}.
}
