% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{heurist_raw_record_edit}
\alias{heurist_raw_record_edit}
\title{Call Heurist Record Edit Directly}
\usage{
heurist_raw_record_edit(session, body = list())
}
\arguments{
\item{session}{A \code{heurist_session}.}

\item{body}{Form body sent to the endpoint.}
}
\value{
Parsed JSON response.
}
\description{
Low-level wrapper around \code{record_edit.php}.
}
