% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{heurist_raw_record_output}
\alias{heurist_raw_record_output}
\title{Call Heurist Record Output Directly}
\usage{
heurist_raw_record_output(session, query = list())
}
\arguments{
\item{session}{A \code{heurist_session}.}

\item{query}{Query parameters passed to the endpoint.}
}
\value{
Parsed JSON response.
}
\description{
Low-level wrapper around \code{record_output.php}.
}
