% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{heurist_rectypes}
\alias{heurist_rectypes}
\title{List Heurist Record Types}
\usage{
heurist_rectypes(session)
}
\arguments{
\item{session}{A \code{heurist_session}.}
}
\value{
A list of Heurist record type objects.
}
\description{
Retrieves record type metadata for the current database.
}
