% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{heurist_restore_change}
\alias{heurist_restore_change}
\title{Restore a Prior Heurist Change}
\usage{
heurist_restore_change(change)
}
\arguments{
\item{change}{A \code{heurist_change} object.}
}
\value{
A \code{heurist_change} object representing the restore operation.
}
\description{
Restores the previous state captured by a \code{heurist_change} object.
}
