% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{heurist_rollback_change}
\alias{heurist_rollback_change}
\title{Roll Back a Heurist Change}
\usage{
heurist_rollback_change(change)
}
\arguments{
\item{change}{A \code{heurist_change} object.}
}
\value{
A \code{heurist_change} object.
}
\description{
Alias for \code{heurist_restore_change()}.
}
