% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heurist-package.R
\docType{package}
\name{heuristR-package}
\alias{heuristR}
\alias{heuristR-package}
\title{heuristR: Session-Oriented Access to Heurist}
\description{
\code{heuristR} provides a session-oriented wrapper around Heurist authentication,
metadata, and record retrieval endpoints. The package is designed to support
safe client-side merge workflows for record updates, rather than relying on
destructive partial save calls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://center-for-archaeology-and-society.github.io/heuristR/}
  \item \url{https://github.com/Center-for-Archaeology-and-Society/heuristR}
  \item \url{https://heuristnetwork.org/}
  \item \url{https://heurist.huma-num.fr/heurist/startup/}
  \item Report bugs at \url{https://github.com/Center-for-Archaeology-and-Society/heuristR/issues}
}

}
\author{
\strong{Maintainer}: Robert J. Bischoff \email{rbischoff@asu.edu}

}
\keyword{internal}
