% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr_import.R
\name{hr_import}
\alias{hr_import}
\title{Read data and paradata from Heurist database}
\usage{
hr_import(data.file, structure.file)
}
\arguments{
\item{data.file}{XML file exported from Heurist including data (fields'values entered in the database) : this XML file can be exported from Heurist via the Publish menu > Export > XML (recommended option). Be careful : if some records are selected in your Heurist session during the export, only data about theses selected records will be exported.}

\item{structure.file}{XML file exported from Heurist describing database structure (vocabularies, fields'caracteristics...) : this XML file can be exported from Heurist via the Design menu > Download > Structure (XML).}
}
\value{
Several dataframes
\itemize{
\item Each Heurist 'record type' as a dataframe : in those dataframes, each row matches a Heurist 'record'
\item A dataframe named "\emph{z.h.tables.fields}" : each row describes a field in an used Heurist 'record type' (type, help text, requirement, repeatability...)
\item A dataframe named "\emph{z.h.vocabularies}" : each row matches a term in an used Heurist vocabulary ; each term is joined to a level (its ranking position in the vocabulary) and attached to its parents'terms
}
}
\description{
This function reads XML files exported from \href{https://heuristnetwork.org/}{Heurist} to build as many dataframes as 'record types' are used in the Heurist database. Each Heurist 'record type' matches a dataframe ; each Heurist 'record' matches a row in a dataframe. Furthermore, this function builds two other dataframes describing fields (type, help text, requirement, repeatability...) and vocabularies.
}
\details{
For more details :
\itemize{
\item a tutorial at \url{https://alietteroux.github.io/heuristr/}
\item the Github repository at \url{https://github.com/alietteroux/heuristr}
}
}
